+++ sentry-rails/spec/sentry/rails/tracing/active_record_subscriber_spec.rb
@@ -30,7 +30,7 @@ RSpec.describe Sentry::Rails::Tracing::ActiveRecordSubscriber, :subscriber do
 
       expect(transport.events.count).to eq(1)
 
-      transaction = transport.events.first.to_hash
+      transaction = transport.events.first.to_h
       expect(transaction[:type]).to eq("transaction")
       expect(transaction[:spans].count).to eq(1)
 
@@ -67,7 +67,7 @@ RSpec.describe Sentry::Rails::Tracing::ActiveRecordSubscriber, :subscriber do
         it "doesn't record query's source location" do
           expect(transport.events.count).to eq(1)
 
-          transaction = transport.events.first.to_hash
+          transaction = transport.events.first.to_h
           expect(transaction[:type]).to eq("transaction")
           expect(transaction[:spans].count).to eq(1)
 
@@ -86,7 +86,7 @@ RSpec.describe Sentry::Rails::Tracing::ActiveRecordSubscriber, :subscriber do
         it "records query's source location" do
           expect(transport.events.count).to eq(1)
 
-          transaction = transport.events.first.to_hash
+          transaction = transport.events.first.to_h
           expect(transaction[:type]).to eq("transaction")
           expect(transaction[:spans].count).to eq(1)
 
@@ -104,7 +104,7 @@ RSpec.describe Sentry::Rails::Tracing::ActiveRecordSubscriber, :subscriber do
         it "doesn't record query's source location" do
           expect(transport.events.count).to eq(1)
 
-          transaction = transport.events.first.to_hash
+          transaction = transport.events.first.to_h
           expect(transaction[:type]).to eq("transaction")
           expect(transaction[:spans].count).to eq(1)
 
@@ -131,7 +131,7 @@ RSpec.describe Sentry::Rails::Tracing::ActiveRecordSubscriber, :subscriber do
 
       expect(transport.events.count).to eq(1)
 
-      transaction = transport.events.first.to_hash
+      transaction = transport.events.first.to_h
       expect(transaction[:type]).to eq("transaction")
       expect(transaction[:spans].count).to eq(2)