+++ sentry-rails/spec/sentry/rails/tracing_spec.rb
@@ -26,8 +26,8 @@ RSpec.describe Sentry::Rails::Tracing, type: :request do
       expect(response).to have_http_status(:internal_server_error)
       expect(transport.events.count).to eq(2)
 
-      event = transport.events.first.to_hash
-      transaction = transport.events.last.to_hash
+      event = transport.events.first.to_h
+      transaction = transport.events.last.to_h
 
       expect(event.dig(:contexts, :trace, :trace_id).length).to eq(32)
       expect(event.dig(:contexts, :trace, :trace_id)).to eq(transaction.dig(:contexts, :trace, :trace_id))
@@ -64,7 +64,7 @@ RSpec.describe Sentry::Rails::Tracing, type: :request do
       expect(response).to have_http_status(:ok)
       expect(transport.events.count).to eq(1)
 
-      transaction = transport.events.last.to_hash
+      transaction = transport.events.last.to_h
 
       expect(transaction[:type]).to eq("transaction")
       expect(transaction.dig(:contexts, :trace, :op)).to eq("http.server")
@@ -117,7 +117,7 @@ RSpec.describe Sentry::Rails::Tracing, type: :request do
 
       it "does not record sensitive params" do
         get "/posts?foo=bar&password=42&secret=baz"
-        transaction = transport.events.last.to_hash
+        transaction = transport.events.last.to_h
 
         params = transaction[:spans][0][:data][:params]
         expect(params["foo"]).to eq("bar")
@@ -139,7 +139,7 @@ RSpec.describe Sentry::Rails::Tracing, type: :request do
 
       it "records all params" do
         get "/posts?foo=bar&password=42&secret=baz"
-        transaction = transport.events.last.to_hash
+        transaction = transport.events.last.to_h
 
         params = transaction[:spans][0][:data][:params]
         expect(params["foo"]).to eq("bar")
@@ -254,7 +254,7 @@ RSpec.describe Sentry::Rails::Tracing, type: :request do
 
       expect(transport.events.count).to eq(3)
 
-      transaction = transport.events.last.to_hash
+      transaction = transport.events.last.to_h
 
       expect(transaction[:type]).to eq("transaction")
       expect(transaction[:transaction]).to eq("PostsController#show")