+++ sentry-ruby/spec/sentry/cron/monitor_config_spec.rb
@@ -45,7 +45,7 @@ RSpec.describe Sentry::Cron::MonitorConfig do
     end
   end
 
-  describe '#to_hash' do
+  describe '#to_h' do
     it 'returns hash with correct attributes for crontab' do
       subject = described_class.from_crontab(
         '5 * * * *',
@@ -54,7 +54,7 @@ RSpec.describe Sentry::Cron::MonitorConfig do
         timezone: 'Europe/Vienna'
       )
 
-      hash = subject.to_hash
+      hash = subject.to_h
       expect(hash).to eq({
         schedule: { type: :crontab, value: '5 * * * *' },
         checkin_margin: 10,
@@ -72,7 +72,7 @@ RSpec.describe Sentry::Cron::MonitorConfig do
         timezone: 'Europe/Vienna'
       )
 
-      hash = subject.to_hash
+      hash = subject.to_h
       expect(hash).to eq({
         schedule: { type: :interval, value: 5, unit: :hour },
         checkin_margin: 10,