+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased (6.0.0)
+
+### Breaking Changes
+
+- Remove `config.async` [#1894](https://github.com/getsentry/sentry-ruby/pull/1894)
+- Migrate from to_hash to to_h ([#2351](https://github.com/getsentry/sentry-ruby/pull/2351))
+
 ## Unreleased
 
 ### Features